// src/app/modules/master-forms/Models/StokistTransportModel.ts
export class StokistTransportModel {
    Mappingid: number = 0;
    BranchId: number = 0;
    CompanyId: number = 0;
    StockistId: number = 0;
    StockistNo: string = "";
    StockistName: string = "";
    Emailid: string = "";
    MobNo: string = "";
    CityCode: string = "";
    LocationId: number = 0;
    TransporterId: number = 0;
    TransporterNo: string = "";
    TransporterName: string = "";
    TransitDays: number = 0;
    Addedby: string = "";
    AddedOn: Date = new Date();
    SupplyTypeId: number = 0;
    MasterName: string = "";
    Action: string = "";
}